- separate to different files - add option to fallback value (instead of undefined) - if 'useKeysAsHeaders' is false, then the body keys come from the header keys - add headerProvider to translate the header keys
