+++ src/utils.ts
@@ -1,0 +1,50 @@
+let hasOwnProperty = Object.prototype.hasOwnProperty;
+let propIsEnumerable = Object.prototype.propertyIsEnumerable;
+
+/**
+ * Convert to Object
+ * @param {any} val
+ */
+function toObject(val: any) {
+    if (val === null || val === undefined) {
+        throw new TypeError('Object.assign cannot be called with null or undefined');
+    }
+
+    return Object(val);
+}
+
+/**
+ * Assign data  to new Object
+ * @param {any}   target
+ * @param {any[]} ...source
+ */
+export function objectAssign(target: any, ...source: any[]) {
+    let from: any;
+    let to = toObject(target);
+    let symbols: any;
+
+    for (let s = 1; s < arguments.length; s++) {
+        from = Object(arguments[s]);
+
+        for (let key in from) {
+            if (hasOwnProperty.call(from, key)) {
+                to[key] = from[key];
+            }
+        }
+
+        if ((<any>Object).getOwnPropertySymbols) {
+            symbols = (<any>Object).getOwnPropertySymbols(from);
+            for (let i = 0; i < symbols.length; i++) {
+                if (propIsEnumerable.call(from, symbols[i])) {
+                    to[symbols[i]] = from[symbols[i]];
+                }
+            }
+        }
+    }
+
+    return to;
+}
+
+export function isFloat(input: any) {
+    return +input === input && (!isFinite(input) || Boolean(input % 1));
+}
